requirements.txt
streamlit>=1.50.0
psycopg2-binary>=2.9.0

